Remove dependency on Document in indenter
