+++ src/Workspaces/Core/Portable/Indentation/AbstractIndentationService.Indenter.cs
@@ -4,6 +4,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Threading;
 using Microsoft.CodeAnalysis.Formatting;
@@ -25,34 +26,33 @@ namespace Microsoft.CodeAnalysis.Indentation
             public readonly TextLine LineToBeIndented;
             public readonly CancellationToken CancellationToken;
 
-            public readonly SyntacticDocument Document;
-            public readonly TSyntaxRoot Root;
             public readonly IEnumerable<AbstractFormattingRule> Rules;
             public readonly BottomUpBaseIndentationFinder Finder;
 
-            private readonly ISyntaxFactsService _syntaxFacts;
+            private readonly ISyntaxFacts _syntaxFacts;
             private readonly int _tabSize;
 
-            public readonly SyntaxTree Tree => Document.SyntaxTree;
-            public readonly SourceText Text => Document.Text;
+            public readonly SyntaxTree Tree;
+            public readonly SourceText Text;
+            public readonly TSyntaxRoot Root;
 
             public readonly ISmartTokenFormatter SmartTokenFormatter;
 
             public Indenter(
                 AbstractIndentationService<TSyntaxRoot> service,
-                SyntacticDocument document,
-                IEnumerable<AbstractFormattingRule> rules,
+                SyntaxTree tree,
+                ImmutableArray<AbstractFormattingRule> rules,
                 IndentationOptions options,
                 TextLine lineToBeIndented,
                 ISmartTokenFormatter smartTokenFormatter,
                 CancellationToken cancellationToken)
             {
-                Document = document;
-
                 _service = service;
-                _syntaxFacts = document.Document.GetRequiredLanguageService<ISyntaxFactsService>();
+                _syntaxFacts = service.SyntaxFacts;
                 Options = options;
-                Root = (TSyntaxRoot)document.Root;
+                Tree = tree;
+                Text = tree.GetText(cancellationToken);
+                Root = (TSyntaxRoot)tree.GetRoot(cancellationToken);
                 LineToBeIndented = lineToBeIndented;
                 _tabSize = options.FormattingOptions.TabSize;
                 SmartTokenFormatter = smartTokenFormatter;
@@ -64,7 +64,7 @@ namespace Microsoft.CodeAnalysis.Indentation
                     _tabSize,
                     options.FormattingOptions.IndentationSize,
                     tokenStream: null,
-                    document.Document.GetRequiredLanguageService<IHeaderFactsService>());
+                    service.HeaderFacts);
             }
 
             public IndentationResult? GetDesiredIndentation(FormattingOptions.IndentStyle indentStyle)
@@ -76,7 +76,7 @@ namespace Microsoft.CodeAnalysis.Indentation
                 // If the user has explicitly set 'block' indentation, or they're in an inactive preprocessor region,
                 // then just do simple block indentation.
                 if (indentStyle == FormattingOptions.IndentStyle.Block ||
-                    _syntaxFacts.IsInInactiveRegion(Document.SyntaxTree, LineToBeIndented.Start, this.CancellationToken))
+                    _syntaxFacts.IsInInactiveRegion(this.Tree, LineToBeIndented.Start, this.CancellationToken))
                 {
                     return GetDesiredBlockIndentation();
                 }
@@ -150,7 +150,7 @@ namespace Microsoft.CodeAnalysis.Indentation
                 // text on it.  We then set our indentation to whatever the indentation of that line was.
                 for (var currentLine = this.LineToBeIndented.LineNumber - 1; currentLine >= 0; currentLine--)
                 {
-                    var line = this.Document.Text.Lines[currentLine];
+                    var line = this.Text.Lines[currentLine];
                     var offset = line.GetFirstNonWhitespaceOffset();
                     if (offset == null)
                         continue;
