+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/CSharp/Extensions/SyntaxTreeExtensions.cs
@@ -29,69 +29,6 @@ namespace Microsoft.CodeAnalysis.CSharp.Extensions
                 syntaxTree.IsInInactiveRegion(position, cancellationToken);
         }
 
-        public static bool IsInInactiveRegion(
-            this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
-        {
-            Contract.ThrowIfNull(syntaxTree);
-
-            // cases:
-            // $ is EOF
-
-            // #if false
-            //    |
-
-            // #if false
-            //    |$
-
-            // #if false
-            // |
-
-            // #if false
-            // |$
-
-            if (syntaxTree.IsPreProcessorKeywordContext(position, cancellationToken))
-            {
-                return false;
-            }
-
-            // The latter two are the hard cases we don't actually have an 
-            // DisabledTextTrivia yet. 
-            var trivia = syntaxTree.GetRoot(cancellationToken).FindTrivia(position, findInsideTrivia: false);
-            if (trivia.Kind() == SyntaxKind.DisabledTextTrivia)
-            {
-                return true;
-            }
-
-            var token = syntaxTree.FindTokenOrEndToken(position, cancellationToken);
-            if (token.Kind() == SyntaxKind.EndOfFileToken)
-            {
-                var triviaList = token.LeadingTrivia;
-                foreach (var triviaTok in triviaList.Reverse())
-                {
-                    if (triviaTok.Span.Contains(position))
-                    {
-                        return false;
-                    }
-
-                    if (triviaTok.Span.End < position)
-                    {
-                        if (!triviaTok.HasStructure)
-                        {
-                            return false;
-                        }
-
-                        var structure = triviaTok.GetStructure();
-                        if (structure is BranchingDirectiveTriviaSyntax branch)
-                        {
-                            return !branch.IsActive || !branch.BranchTaken;
-                        }
-                    }
-                }
-            }
-
-            return false;
-        }
-
         public static bool IsInPartiallyWrittenGeneric(
             this SyntaxTree syntaxTree, int position, CancellationToken cancellationToken)
         {
